Move the cluster name from ConfigWriter to SimpleStore.
Reason: if left ConfigWriter, nodes don't know to which cluster they belong.
This will bite us later when we'll revisit node join operation.

Cons: we lose the cluster name from the config file, which means a
standalone "cluster.data" is a little bit harder to analyze. But I have
a plan for this which will follow shortly.

Implementing this will also allow us to cleanly fix "gnt-cluster
getmaster".

Reviewed-by: imsnah