Another documentation update.
This removes the section on packages from installing.sgml (until we have
packages) and adds/corrects informations about using the startup script.

It also adds notes about the ``--master-netdev`` option, which one needs
to use if the node doesn't have an interface named xen-br0 (or if it's
not the interface wanted for external access).