Documentation fix for missing dependencies
I was getting into trouble with gnt-backup and discovered reading the
code of the debian-etch os image and the error logs that:
     1 - Debian os image shipped by ganeti depends on the availability
of dump and restore.

Based on that I had included in the section of "Installing the
Operating System support packages" the dependency of dump/restore and
the command to install dump/restore.

Author: Leonardo Rodrigues de Mello <l@lmello.eu.org>
Reviewed-by: iustinp