Make instance start/stop skippable at burnin time
Even though burnin was born just to do that test it now contains a lot more
things one might try, so it makes sense to make instance start/stop optional
too.

This creates a burnin that at the bare minimum tests instance create and
remove, if all the --no options are specified, but usually does a lot more.

Reviewed-by: iustinp